[FIX] Air Dash logic rewrite
Placing the parts of logic where we're changing the airdash and consolidating it within the air dash state.  Since the calculation of the air dash trajectory will be in the very first frame prior to the actual execution of the dash.  This should be more accurate to what we want

I hope it works